chore: carousel 명칭 변경
좀 더 범용적으로 사용되는 컴포넌트이기 때문에 prefix 제거
